Add `useFullHeight` hook
